chore(cleanup): cleanup folder structure
