+++ lib/constants/constants.ts
@@ -1,0 +1,15 @@
+
+const USER_STORAGE_KEY = 'iota-user'
+const API_BASE_URL = "http://localhost:1337/api";
+const FORM_PARENT_STYLES = {
+    border: "0.1px solid rgba(0,0,0,0.1)",
+    borderRadius: 15,
+    backgroundColor: "white",
+    overflow: "hidden",
+};
+
+export default {
+    USER_STORAGE_KEY,
+    API_BASE_URL,
+    FORM_PARENT_STYLES
+}
