+++ lib/constants/itemCreationSteps.ts
@@ -1,0 +1,32 @@
+export default [
+    {
+        title: 'General',
+        content: 'First-content',
+    },
+    {
+        title: 'Specifics',
+        content: 'Second-content',
+    },
+    {
+        title: 'Logistics',
+        content: 'Last-content',
+    },
+];
+
+/**
+ * 1. General Information
+ * name - text input
+ * description - text input
+ * category - dropdwon select
+ * quantity - number input
+ * colour - color picker
+ * 
+ * 2. Logistics Information
+ * supplier - text input
+ * manufacturer - text input
+ * handling - dropdwon select
+ * 
+ * 3. Specific Information
+ * dimensions - number inputs then dropdown for units
+ * weight - number inputs then dropdown for units
+ */
