+++ lib/users.ts
@@ -1,11 +1,11 @@
 import { get } from 'lodash';
 
-import { API_BASE_URL, HTTP_ERRORS } from '../lib/constants';
+import { GENERAL_CONSTANTS, HTTP_ERRORS } from './constants';
 import { UserDetails, UserLoginDetails, LoggedInUserDetails } from '../lib/types';
 
 export async function signupUser(values: UserDetails): Promise<LoggedInUserDetails> {
     try {
-        const res = await fetch(`${API_BASE_URL}/auth/local/register`, {
+        const res = await fetch(`${GENERAL_CONSTANTS.API_BASE_URL}/auth/local/register`, {
             method: 'POST',
             headers: { 'Content-Type': 'application/json' },
             body: JSON.stringify(values)
@@ -26,7 +26,7 @@ export async function signupUser(values: UserDetails): Promise<LoggedInUserDetai
 
 export async function loginUser(details: UserLoginDetails): Promise<LoggedInUserDetails> {
     try {
-        const res = await fetch(`${API_BASE_URL}/auth/local`, {
+        const res = await fetch(`${GENERAL_CONSTANTS.API_BASE_URL}/auth/local`, {
             method: 'POST',
             headers: { 'Content-Type': 'application/json' },
             body: JSON.stringify(details)
